+++ packages/frontend/synthesizer/core/src/synthesizer/handlers/arithmeticManager.ts
@@ -121,7 +121,7 @@ export class ArithmeticManager {
       finalInPts = [selectorPt, ...finalInPts];
     }
 
-    if (name === 'SHL' || name === 'SHR' || name === 'SAR') {
+    if (name === 'SHR' || name === 'SAR') {
       if (inPts[0] !== undefined && inPts[0].value > 255n) {
         throw new Error(
           `Synthesizer: Operation ${name} has a shift value greater than 255. Adjust ${subcircuitName} subcircuit in qap-compiler.`,